:recycle: use service to check if sessions has started
