+++ api/lib/domain/usecases/create-sessions.js
@@ -18,7 +18,6 @@ module.exports = async function createSessions({
   certificationCpfCityRepository,
   certificationCandidateRepository,
   complementaryCertificationRepository,
-  certificationCourseRepository,
 }) {
   const { name: certificationCenter, isSco } = await certificationCenterRepository.get(certificationCenterId);
 
@@ -35,9 +34,6 @@ module.exports = async function createSessions({
       });
 
       if (sessionId) {
-        if (await _isSessionStarted({ certificationCourseRepository, sessionId })) {
-          throw new UnprocessableEntityError("Impossible d'ajouter un candidat à une session qui a déjà commencé.");
-        }
         await _deleteExistingCandidatesInSession({ certificationCandidateRepository, sessionId, domainTransaction });
       }
 
@@ -94,13 +90,6 @@ function _hasDuplicateCertificationCandidates(certificationCandidates) {
   );
 
   return uniqCertificationCandidates.size < certificationCandidates.length;
-}
-
-async function _isSessionStarted({ certificationCourseRepository, sessionId }) {
-  const foundCertificationCourses = await certificationCourseRepository.findCertificationCoursesBySessionId({
-    sessionId,
-  });
-  return foundCertificationCourses.length > 0;
 }
 
 async function _deleteExistingCandidatesInSession({ certificationCandidateRepository, sessionId, domainTransaction }) {
