+++ api/tests/unit/domain/usecases/create-sessions_test.js
@@ -1,7 +1,6 @@
 const { domainBuilder, expect, catchErr, sinon } = require('../../../test-helper');
 const createSessions = require('../../../../lib/domain/usecases/create-sessions');
 const { EntityValidationError, InvalidCertificationCandidate } = require('../../../../lib/domain/errors');
-const { UnprocessableEntityError } = require('../../../../lib/application/http-errors');
 const sessionCodeService = require('../../../../lib/domain/services/session-code-service');
 const Session = require('../../../../lib/domain/models/Session');
 const certificationCpfService = require('../../../../lib/domain/services/certification-cpf-service');
@@ -15,7 +14,6 @@ describe('Unit | UseCase | create-sessions', function () {
   let certificationCenterName;
   let certificationCenter;
   let certificationCenterRepository;
-  let certificationCourseRepository;
   let certificationCandidateRepository;
   let complementaryCertificationRepository;
   let sessionRepository;
@@ -29,7 +27,6 @@ describe('Unit | UseCase | create-sessions', function () {
       name: certificationCenterName,
     });
     certificationCenterRepository = { get: sinon.stub() };
-    certificationCourseRepository = { findCertificationCoursesBySessionId: sinon.stub() };
     certificationCandidateRepository = { saveInSession: sinon.stub(), deleteBySessionId: sinon.stub() };
     complementaryCertificationRepository = { getByLabel: sinon.stub() };
     sessionRepository = { save: sinon.stub(), isSessionExisting: sinon.stub() };
@@ -93,35 +90,6 @@ describe('Unit | UseCase | create-sessions', function () {
       });
 
       context('when there is only sessionId and candidate information', function () {
-        context('when a session has already started', function () {
-          it('should throw', async function () {
-            const candidate = createValidCandidateData(1);
-            const sessions = [
-              {
-                sessionId: 1234,
-                certificationCandidates: [candidate],
-              },
-            ];
-
-            certificationCourseRepository.findCertificationCoursesBySessionId
-              .withArgs({ sessionId: sessions[0].sessionId })
-              .resolves([domainBuilder.buildCertificationCourse({ sessionId: 1234 })]);
-
-            // when
-            const error = await catchErr(createSessions)({
-              sessions,
-              certificationCenterId,
-              certificationCenterRepository,
-              certificationCourseRepository,
-              sessionRepository,
-            });
-
-            // then
-            expect(error).to.be.instanceOf(UnprocessableEntityError);
-            expect(error.message).to.equal("Impossible d'ajouter un candidat à une session qui a déjà commencé.");
-          });
-        });
-
         it('should only save candidate in session', async function () {
           // given
           const candidate1 = createValidCandidateData(1);
@@ -137,7 +105,6 @@ describe('Unit | UseCase | create-sessions', function () {
           const cpfBirthInformationValidation2 = CpfBirthInformationValidation.success({ ...candidate2 });
           certificationCpfService.getBirthInformation.onCall(0).resolves(cpfBirthInformationValidation1);
           certificationCpfService.getBirthInformation.onCall(1).resolves(cpfBirthInformationValidation2);
-          certificationCourseRepository.findCertificationCoursesBySessionId.resolves([]);
 
           const domainTransaction = Symbol('trx');
           sinon.stub(DomainTransaction, 'execute').callsFake((lambda) => lambda(domainTransaction));
@@ -148,7 +115,6 @@ describe('Unit | UseCase | create-sessions', function () {
             certificationCenterId,
             certificationCenterRepository,
             certificationCandidateRepository,
-            certificationCourseRepository,
             sessionRepository,
           });
